refactor(StorageSize): Organzia storage file and code
